Final Project/backend/requirements.txt
Flask==3.0.0
requests
firebase_admin
google-cloud-pubsub
flask_cors
uuid
gunicorn
datetime